PN532 error causes the device to show light orange LED

+++ Source/Tappt/Tap/ITap.h
@@ -6,7 +6,7 @@
 class ITap {
 public:
   virtual String GetId() = 0;
-  virtual void AddToFlowCount(uint8_t pulses) = 0;
+  virtual void AddToFlowCount(uint pulses) = 0;
   virtual void SetAuthToken(String authenticationKey) = 0;
   virtual ~ITap() {}
 };
